# Tweak sdk tests so they run faster