The test utility `_create_start_and_end_span()` creates a new Resource each time it creates a span. This is quite slow because Resource.create() iterates through entry points every time it's called. This change requires that a Resource be passed in to `_create_start_and_end_span` instead.

Tests that take > 1 second:

BEFORE:

145.20s call     opentelemetry-sdk/tests/trace/export/test_export.py::TestBatchSpanProcessor::test_flush_from_multiple_threads
4.68s call     opentelemetry-sdk/tests/trace/export/test_export.py::TestBatchSpanProcessor::test_batch_span_processor_many_spans
3.12s call     opentelemetry-sdk/tests/trace/export/test_export.py::TestBatchSpanProcessor::test_batch_span_processor_lossless
2.02s call     opentelemetry-sdk/tests/metrics/test_metrics.py::TestDuplicateInstrumentAggregateData::test_duplicate_instrument_aggregate_data
2.01s call     opentelemetry-sdk/tests/metrics/test_measurement_consumer.py::TestSynchronousMeasurementConsumer::test_collect_deadline
1.79s call     opentelemetry-sdk/tests/metrics/exponential_histogram/test_exponential_bucket_histogram_aggregation.py::TestExponentialBucketHistogramAggregation::test_ascending_sequence
1.72s call     opentelemetry-sdk/tests/metrics/test_metric_reader_storage.py::TestMetricReaderStorage::test_race_concurrent_measurements
1.49s call     opentelemetry-sdk/tests/trace/export/test_export.py::TestBatchSpanProcessor::test_batch_span_processor_fork
1.13s call     opentelemetry-sdk/tests/logs/test_export.py::TestBatchLogRecordProcessor::test_batch_log_record_processor_fork
1.10s call     opentelemetry-sdk/tests/performance/benchmarks/metrics/test_benchmark_metrics.py::test_counter_add[0-delta]
1.02s call     opentelemetry-sdk/tests/metrics/integration_test/test_time_align.py::TestTimeAlign::test_time_align_cumulative
1.00s call     opentelemetry-sdk/tests/metrics/test_measurement_consumer.py::TestSynchronousMeasurementConsumer::test_collect_timeout

AFTER:

2.01s call     opentelemetry-sdk/tests/metrics/test_measurement_consumer.py::TestSynchronousMeasurementConsumer::test_collect_deadline
1.73s call     opentelemetry-sdk/tests/metrics/exponential_histogram/test_exponential_bucket_histogram_aggregation.py::TestExponentialBucketHistogramAggregation::test_ascending_sequence
1.37s call     opentelemetry-sdk/tests/metrics/test_metric_reader_storage.py::TestMetricReaderStorage::test_race_concurrent_measurements
1.36s call     opentelemetry-sdk/tests/trace/export/test_export.py::TestBatchSpanProcessor::test_batch_span_processor_fork
1.10s call     opentelemetry-sdk/tests/logs/test_export.py::TestBatchLogRecordProcessor::test_batch_log_record_processor_fork
1.09s call     opentelemetry-sdk/tests/performance/benchmarks/metrics/test_benchmark_metrics.py::test_counter_add[0-delta]
1.03s call     opentelemetry-sdk/tests/trace/export/test_export.py::TestBatchSpanProcessor::test_batch_span_processor_lossless
1.01s call     opentelemetry-sdk/tests/metrics/integration_test/test_time_align.py::TestTimeAlign::test_time_align_cumulative
1.00s call     opentelemetry-sdk/tests/metrics/test_measurement_consumer.py::TestSynchronousMeasurementConsumer::test_collect_timeout