@pmcollins 2023-11-06
> This looks promising, but I don't see the before and after times being comparable, it seems like there are items in the before list that don't show up in the after list.

Thanks -- to address your comment, it's because these are just the tests that take > 1 second. If a test shows up in the first list and doesn't show up in the second list it's because after this change the test is no longer slower than 1 second.